feat(brain-global-save-button): adds tunings to global Brain mixin state manager
